test/fixtures/wpt/resource-timing/clear-resource-timings.html

<!DOCTYPE HTML>
<html>
<head>
<meta charset="utf-8" />
<title>This test validates the functionality of clearResourceTimings method
in resource timing.</title>
<link rel="author" title="Intel" href="http://www.intel.com/" />
<link rel="help"
 href="https://www.w3.org/TR/resource-timing-2/#dom-performance-clearresourcetimings">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
test(() => {
  assert_equals(performance.getEntriesByType("resource").length, 2,
                "Resource timing entries exist");
  performance.clearResourceTimings();
  assert_equals(performance.getEntriesByType("resource").length, 0,
                "Resource timing entries are cleared");
}, "Test that clearResourceTimings() clears the performance timeline buffer");
</script>
</head>
</html>
